Fix warm capacity helper build tags
